+++ internal/gateway/errors.go
@@ -7,4 +7,5 @@ var (
 	ErrDoesNotExist               = errors.New("gateway does not exist")
 	ErrAlreadyExists              = errors.New("gateway already exists")
 	ErrInvalidAggregationInterval = errors.New("invalid aggregation interval")
+	ErrInvalidName                = errors.New("invalid gateway name")
 )